# apps/server/openapi.json
{
  "components": {
    "schemas": {
...
      "AnnotationOut": {
        "additionalProperties": false,
        "properties": {
          "authorId": {
            "title": "Authorid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "scope": {
            "enum": [
              "private",
              "project"
            ],
            "title": "Scope",
            "type": "string"
          },
          "scoreId": {
            "title": "Scoreid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scoreId",
          "authorId",
          "scope",
          "revision",
          "data",
          "createdAt",
          "updatedAt"
        ],
        "title": "AnnotationOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API and WebSocket clock/transport gateway. Clients expand tempo-map beats locally; the server only agrees on revision, anchor, and absolute server time.",
    "title": "FeelMyRythm API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/annotations/{annotation_id}": {
      "delete": {
        "operationId": "delete_annotation_api_annotations__annotation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "annotation_id",
            "required": true,
            "schema": {
              "title": "Annotation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Annotation",
        "tags": [
          "scores"
        ]
      },
      "get": {
        "operationId": "get_annotation_api_annotations__annotation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "annotation_id",
            "required": true,
            "schema": {
              "title": "Annotation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Annotation",
        "tags": [
          "scores"
        ]
      },
      "put": {
        "operationId": "update_annotation_api_annotations__annotation_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "annotation_id",
            "required": true,
            "schema": {
              "title": "Annotation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Annotation",
        "tags": [
          "scores"
        ]
      }
    },
...
    "/api/repertoire/{repertoire_id}/annotations": {
      "get": {
        "operationId": "list_repertoire_annotations_api_repertoire__repertoire_id__annotations_get",
        "parameters": [
          {
            "in": "path",
            "name": "repertoire_id",
            "required": true,
            "schema": {
              "title": "Repertoire Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnnotationOut"
                  },
                  "title": "Response List Repertoire Annotations Api Repertoire  Repertoire Id  Annotations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Repertoire Annotations",
        "tags": [
          "scores"
        ]
      }
    },
...
    "/api/repertoire/{repertoire_id}/scores": {
      "get": {
        "operationId": "list_scores_api_repertoire__repertoire_id__scores_get",
        "parameters": [
          {
            "in": "path",
            "name": "repertoire_id",
            "required": true,
            "schema": {
              "title": "Repertoire Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScoreOut"
                  },
                  "title": "Response List Scores Api Repertoire  Repertoire Id  Scores Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Scores",
        "tags": [
          "scores"
        ]
      }
    },
    "/api/repertoire/{repertoire_id}/scores/presign": {
      "post": {
        "operationId": "presign_score_api_repertoire__repertoire_id__scores_presign_post",
        "parameters": [
          {
            "in": "path",
            "name": "repertoire_id",
            "required": true,
            "schema": {
              "title": "Repertoire Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScorePresignIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTargetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Presign Score",
        "tags": [
          "scores"
        ]
      }
    },
...
    "/api/scores/{score_id}": {
      "delete": {
        "operationId": "delete_score_api_scores__score_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "score_id",
            "required": true,
            "schema": {
              "title": "Score Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Score",
        "tags": [
          "scores"
        ]
      },
      "get": {
        "operationId": "get_score_api_scores__score_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "score_id",
            "required": true,
            "schema": {
              "title": "Score Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Score",
        "tags": [
          "scores"
        ]
      },
      "patch": {
        "operationId": "update_score_api_scores__score_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "score_id",
            "required": true,
            "schema": {
              "title": "Score Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Score",
        "tags": [
          "scores"
        ]
      }
    },
    "/api/scores/{score_id}/annotations": {
      "get": {
        "operationId": "list_annotations_api_scores__score_id__annotations_get",
        "parameters": [
          {
            "in": "path",
            "name": "score_id",
            "required": true,
            "schema": {
              "title": "Score Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnnotationOut"
                  },
                  "title": "Response List Annotations Api Scores  Score Id  Annotations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Annotations",
        "tags": [
          "scores"
        ]
      },
      "post": {
        "operationId": "create_annotation_api_scores__score_id__annotations_post",
        "parameters": [
          {
            "in": "path",
            "name": "score_id",
            "required": true,
            "schema": {
              "title": "Score Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Annotation",
        "tags": [
          "scores"
        ]
      }
    },
...
  }
}